Crate fasten_mutate

Mutates reads. There is no mutation model; only randomness.

Examples

cat testdata/four_reads.fastq | fasten_mutate > out.fastq

Usage

 
fasten_mutate: Introduces point mutations randomly. There is no
evolutionary model; multiple hits are allowed. Therefore,
the number of SNPs through --snps is an upper
limit.
 
Usage: fasten_mutate [-h] [-n INT] [-p] [--verbose] [--version] [-s INT] [-m]
 
Options:
    -h, --help          Print this help menu.
    -n, --numcpus INT   Number of CPUs (default: 1)
    -p, --paired-end    The input reads are interleaved paired-end
        --verbose       Print more status messages
        --version       Print the version of Fasten and exit
    -s, --snps INT      Maximum number of SNPs (point mutations) to include
                        per read.
    -m, --mark          lowercase all reads but uppercase the SNPs (not yet
                        implemented)
 

Functions

main 🔒

mutate 🔒

Mutate a str of a sequence of nucleotides using the nucleotides in a vector nts. This function does not use any kind of mutation model and will choose random positions to replace with random nucleotides.